+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **fix**: check command code when compiling
 - **refactor**: use static class member functions
 - **fix**: data folder is incorrect in file-only mode
+- **fix**: hide the warning message when the checker or interactor is an
+  executable file
 
 ## 0.1.10